+++ src/main/java/com/example/autogarage/repsitory/StockRepository.java
@@ -3,5 +3,5 @@ package com.example.autogarage.repsitory;
 import com.example.autogarage.model.Stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockRepository extends JpaRepository<Stock,Long> {
+public interface StockRepository extends JpaRepository<Stock,Integer> {
 }
